+++ TheatreProject.WebApp/Models/DTOs/CartHeaderDto.cs
@@ -5,8 +5,8 @@ public class CartHeaderDto
     public Guid Id { get; set; }
     public string? UserId { get; set; }
     public string? CouponCode { get; set; }
-    public double DiscountTotal { get; set; }
-    public double GrandTotal { get; set; }
+    public decimal DiscountTotal { get; set; }
+    public decimal GrandTotal { get; set; }
     public string? Email { get; set; }
     public string? FirstName { get; set; }
     public string? LastName { get; set; }
