+++ TheatreProject.WebApp/Models/DTOs/PerformanceStatisticsDto.cs
@@ -5,5 +5,5 @@ public class PerformanceStatisticsDto
     public int TotalBookings { get; set; }
     public decimal TotalRevenue { get; set; }
     public int AvailableSeats { get; set; }
-    public double OccupancyRate { get; set; }
+    public decimal OccupancyRate { get; set; }
 }
